chore: remove unused code, directories, and tests
