+++ app/views/layouts/application.phtml
@@ -60,12 +60,6 @@
                     <a class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100" href="<?= route('diets.index') ?>">Minhas dietas</a>
                 </li>
                 <li>
-                    <a class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100" href="<?= route('reinforce.problems.supported') ?>">Alimentos adicionados</a>
-                </li>
-                <li>
-                    <a class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100" href="<?= route('reinforce.problems') ?>">Todas as dietas</a>
-                </li>
-                <li>
                     <a class="flex items-center block gap-2 px-4 py-2 text-sm text-gray-700 hover:bg-gray-100" href="<?= route('profile.show') ?>">
                         <i class="bi bi-person-fill text-xl text-black"></i>
                         <span class="font-medium">Perfil</span>
