--- app/views/problems/new.phtml
@@ -1,14 +1,0 @@
-<form action="<?= route('problems.create') ?>" method="POST" novalidate>
-    <div class="mb-4">
-        <label for="problem_title" class="block text-sm font-medium text-gray-700 mb-1">Título:</label>
-        <input type="text" value="<?= $problem->title ?>" name="problem[title]" id="problem_title" required
-            class="w-full border border-gray-300 rounded-md px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500">
-        <?php if ($problem->errors('title')) : ?>
-            <p class="text-red-600 text-xs mt-1"><?= $problem->errors('title') ?></p>
-        <?php endif; ?>
-    </div>
-
-    <button type="submit" class="px-4 py-2 bg-blue-600 text-white text-sm rounded-md hover:bg-blue-700">
-        Registrar
-    </button>
-</form>
