--- app/views/problems/show.phtml
@@ -1,3 +1,0 @@
-<div class="bg-white border border-gray-200 rounded-lg shadow-sm p-6">
-    <?= $problem->title ?>
-</div>
