--- app/views/reinforce_problems/index.phtml
@@ -1,49 +1,0 @@
-<div class="flex gap-4 mb-4">
-    <div class="flex max-w-md">
-        <input type="text" class="flex-1 border border-gray-300 rounded-l-md px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500" placeholder="Procurar alimentos">
-        <button class="px-4 py-2 text-sm border border-l-0 border-gray-300 rounded-r-md hover:bg-gray-100 text-gray-700">
-            Pesquisar...
-        </button>
-    </div>
-</div>
-
-<table class="w-full text-sm text-left text-gray-700">
-    <thead class="border-b border-gray-200 text-gray-500 uppercase text-xs">
-        <tr>
-            <th class="py-3 pr-4">ID</th>
-            <th class="py-3 pr-4">Título</th>
-            <th class="py-3 pr-4">Registrado por</th>
-            <th class="py-3">Apoios</th>
-        </tr>
-    </thead>
-    <tbody class="divide-y divide-gray-100">
-        <?php foreach ($problems as $problem) : ?>
-            <tr class="hover:bg-gray-50">
-                <td class="py-3 pr-4">#<?= $problem->id ?></td>
-                <td class="py-3 pr-4"><?= $problem->title ?></td>
-                <td class="py-3 pr-4"><?= $problem->user->name ?></td>
-                <td class="py-3">
-                    <div class="flex items-center gap-3">
-                        <?php if ($problem->isSupportedByUser($this->current_user)) : ?>
-                            <form action="<?= route('reinforce.problems.stopped-supporting', ['id' => $problem->id]) ?>" method="POST" class="m-0">
-                                <button class="px-3 py-1 text-sm border border-red-500 text-red-500 rounded-md hover:bg-red-50 cursor-pointer">
-                                    <i class="bi bi-hand-thumbs-down"></i> Desapoiar
-                                </button>
-                            </form>
-                        <?php else : ?>
-                            <form action="<?= route('reinforce.problems.create', ['id' => $problem->id]) ?>" method="POST" class="m-0">
-                                <button class="px-3 py-1 text-sm border border-blue-600 text-blue-600 rounded-md hover:bg-blue-50 cursor-pointer">
-                                    <i class="bi bi-hand-thumbs-up"></i> Apoiar
-                                </button>
-                            </form>
-                        <?php endif; ?>
-
-                        <span class="text-gray-500"><?= $problem->reinforcedByUsers()->count() ?> Apoiadores</span>
-                    </div>
-                </td>
-            </tr>
-        <?php endforeach; ?>
-    </tbody>
-</table>
-
-<?php $paginator->renderPagesNavigation(); ?>
