--- app/views/reinforce_problems/supported.phtml
@@ -1,18 +1,0 @@
-<table class="w-full text-sm text-left text-gray-700">
-    <thead class="border-b border-gray-200 text-gray-500 uppercase text-xs">
-        <tr>
-            <th class="py-3 pr-4">ID</th>
-            <th class="py-3 pr-4">Título</th>
-            <th class="py-3 pr-4">Registrado por</th>
-        </tr>
-    </thead>
-    <tbody class="divide-y divide-gray-100">
-        <?php foreach ($problems as $problem) : ?>
-            <tr class="hover:bg-gray-50">
-                <td class="py-3 pr-4">#<?= $problem->id ?></td>
-                <td class="py-3 pr-4"><?= $problem->title ?></td>
-                <td class="py-3 pr-4"><?= $problem->user->name ?></td>
-            </tr>
-        <?php endforeach; ?>
-    </tbody>
-</table>
