+++ database/Populate/populate.php
@@ -3,10 +3,8 @@
 require __DIR__ . '/../../config/bootstrap.php';
 
 use Core\Database\Database;
-use Database\Populate\ProblemsPopulate;
 use Database\Populate\UsersPopulate;
 
 Database::migrate();
 
 UsersPopulate::populate();
-ProblemsPopulate::populate();
