+++ tests/Integration/Access/DietsAccessTest.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Tests\Integration\Access;
+
+use GuzzleHttp\Client;
+use PHPUnit\Framework\TestCase;
+
+class DietsAccessTest extends TestCase
+{
+    private Client $client;
+
+    public function setup(): void
+    {
+        parent::setUp();
+        $this->client = new Client([
+            'allow_redirects' => false, // Disable following redirects
+            'base_uri' => 'http://web:8080'
+        ]);
+    }
+}
